chore: Update eslint configuration to include rules for no-console and no-unused-vars
This commit updates the .eslintrc.json file to include the "no-console" and "no-unused-vars" rules in the parserOptions. These rules will now warn when console statements or unused variables are detected in the code.

Note: This commit message follows the established convention of starting with a verb in the imperative form and using the "chore" type for miscellaneous changes.